Move OptionT from std to data
